Pass signals through to the child process
When the `appsignal-wrap` process receives a signal that is commonly
used for inter-process communication or to signal termination to the
process, instead of having it trigger its default behaviour, trap it
and pass it on to the child process.

So, for example, instead of having `SIGINT` finish `appsignal-wrap`,
triggering Tokio to kill the child process, forward this `SIGINT`
to the child process, allowing it to perform whichever shutdown it
needs, then having `appsignal-wrap` finish shortly after.

After the child process has finished, `appsignal-wrap` may still be
awaiting some deliveries of logs and check-ins. Since we set up
signal handlers earlier to forward events to the child process, we
must continue to listen to those signal handlers while
`appsignal-wrap` is finishing its tasks, so that the wrapper itself
is cancellable by ordinary signals both before and after the child
process runs.