Exit with the child process' exit code
If the child process exits with an exit code, exit with that exit
code.

If the child process does not exit with an exit code (meaning that
its exit was caused by a signal) then add 128 to the code of the
signal that caused the exit and use that as the exit code. This is
how shells represent signal exit statuses.