Add digests to cron check-ins and logs
Add a digest when reporting start and finish cron check-ins. This
ensures that the finish digest closes the cron check-in occurrence
that was opened by its start check-in event, instead of closing
some other digest-less cron check-in event.

Also add a digest to the logs as an attribute. This allows logs
from multiple, potentially simultaneous invocations of `appsignal-wrap`
to be correlated per execution. The digest matches the one from the
cron check-ins, allowing the logs to be correlated with the events
in a specific cron check-in occurrence.