Exit child process if parent process dies
In the event of `appsignal-wrap` crashing, the child process it
spawns should be killed. This prevents a scenario where a process
monitor, such as `systemd`, restarts `appsignal-wrap` in a loop,
with it crashing and leaving many copies of the child process
running as zombie processes.

This is only implemented for Linux, as macOS requires a much more
complicated [setup][pthread-kqueue-horror-show] that I could not
successfully reimplement in Rust.

[pthread-kqueue-horror-show]: https://gitlab.com/nbdkit/nbdkit/-/blob/f80723fc76724e94bdf0e41a13b1c9280304d957/common/utils/exit-with-parent.c#L96-152