Improve test coverage
Add tests for the requests that can be produced by the log, cron and
heartbeat configs, and for the logic that emits those configs from
the `Cli` struct.

Refactor some things to make those tests possible -- add a `Timestamp`
trait that produces timestamps, which will be useful to implement
monotonically increasing timestamps as described in #7, but it is
also useful for having the tests be timestamp-independent, and
aliasing over the different kinds of timestamps that our public API
endpoints expect.