Restore default (ignore) Rust SIGPIPE behavior
Remove `reset_sigpipe` and other code that handles `SIGPIPE` in any
way, such as ignoring it or forwarding it to the child process.

This restores the default behaviour of the Rust standard library,
which is to always ignore the signal. This behaviour will be inherited
by the child process.

See #12 for context.