Report non-success exits as errors
When the child process exits due to a signal, or exits with a
non-zero exit code, report this as an error to AppSignal. Closes #2.

This commit implements this functionality as opt-in, with an `--error`
flag. This will probably be changed to opt-out once #5 is implemented,
which will provide a name to group by.

Set the hostname, digest, and exit information as tags on the error
sample.