Add last lines of output to error message
When `--error` is set, use the last lines of standard output and
standard error as the error message.

This honors the `--no-stdout` and `--no-stderr` options -- if set,
standard output or standard error won't be used for either logs or
error messages.

A separate message processing loop is used to keep the last ten
lines of output and use them as the message.