Add installation script to release
Currently, the installation script referenced in the README.md is
the one currently available in the `main` branch. This means that,
if we were to change the installation script in the `main` branch,
in concordance with a different change that was not yet released,
the README would point to an installation script that might not
work correctly with the latest release.

To fix this, make the installation script part of the release
itself.