Actually publish macOS releases
A [previous PR][oops] attempted to build macOS releases from an
Ubuntu runner, which cannot possibly work.

[oops]: https://github.com/appsignal/appsignal-wrap/pull/26

I then [attempted to do the whole build in a macOS runner][whoopsie],
which also didn't work, as running Docker on macOS requires Docker
Desktop.

[whoopsie]: https://github.com/appsignal/appsignal-wrap/compare/9907babe8f0ca41262b56595fb6841a16f5ed7e6...2ed30706a2bb4dd39d348675628ec71df6ebec46

This commit follows [the agent's approach][:brain:], using a build
matrix to build macOS builds on a macOS runner and Linux builds on
an Ubuntu runner, then publishing it from an Ubuntu runner in a
separate job.

[:brain:]: https://github.com/appsignal/appsignal-agent/blob/96b5b6721804e7e15546ed6faf86348b5fa01c13/.github/workflows/publish_release.yml#L81-L100